Refactor cleanup
